Move each view to its own state
