+++ features/mpmoviedetails/src/main/java/com/jpp/mpmoviedetails/MovieDetailActionViewState.kt
@@ -14,7 +14,7 @@ internal data class MovieDetailActionViewState(
     val rateButtonState: ActionButtonState = ActionButtonState(),
     val watchListButtonState: ActionButtonState = ActionButtonState(),
     val favoriteButtonState: ActionButtonState = ActionButtonState(),
-    val errorState: ErrorState = ErrorState.None,
+    val errorState: ActionErrorViewState = ActionErrorViewState.None,
     val animate: Boolean = false,
     val expanded: Boolean = false
 ) {
@@ -30,7 +30,7 @@ internal data class MovieDetailActionViewState(
             rateButtonState = rateButtonState.asVisible(),
             watchListButtonState = watchListButtonState,
             favoriteButtonState = favoriteButtonState,
-            errorState = ErrorState.None,
+            errorState = ActionErrorViewState.None,
             animate = false,
             expanded = false
         )
@@ -44,7 +44,7 @@ internal data class MovieDetailActionViewState(
             rateButtonState = rateButtonState.asInVisible(),
             watchListButtonState = watchListButtonState.asInVisible(),
             favoriteButtonState = favoriteButtonState.asInVisible(),
-            errorState = ErrorState.UnknownError,
+            errorState = ActionErrorViewState.UnknownError,
             animate = false,
             expanded = false
         )
@@ -70,7 +70,7 @@ internal data class MovieDetailActionViewState(
             rateButtonState = rateButtonState.asInVisible(),
             watchListButtonState = watchListButtonState.asInVisible(),
             favoriteButtonState = favoriteButtonState.asInVisible(),
-            errorState = ErrorState.UserNotLogged,
+            errorState = ActionErrorViewState.UserNotLogged,
             animate = true,
             expanded = false
         )
@@ -80,67 +80,4 @@ internal data class MovieDetailActionViewState(
         fun showLoading(): MovieDetailActionViewState =
             MovieDetailActionViewState(loadingVisibility = View.VISIBLE)
     }
-}
-
-/**
- * Represents the state of the action buttons show to the user.
- */
-internal data class ActionButtonState(
-    val visibility: Int = View.INVISIBLE,
-    val animateLoading: Boolean = false,
-    val asFilled: Boolean = false,
-    val asClickable: Boolean = false
-) {
-    fun asVisible(): ActionButtonState {
-        return copy(
-            visibility = View.VISIBLE,
-            asClickable = true
-        )
-    }
-
-    fun asInVisible(): ActionButtonState {
-        return copy(visibility = View.INVISIBLE)
-    }
-
-    fun asFilled(): ActionButtonState {
-        return copy(
-            visibility = View.VISIBLE,
-            asFilled = true,
-            asClickable = true,
-            animateLoading = false
-        )
-    }
-
-    fun asEmpty(): ActionButtonState {
-        return copy(
-            visibility = View.VISIBLE,
-            asFilled = false,
-            asClickable = true,
-            animateLoading = false
-        )
-    }
-
-    fun asLoading(): ActionButtonState {
-        return copy(
-            animateLoading = true,
-            asClickable = false
-        )
-    }
-
-    fun flipState(): ActionButtonState {
-        return copy(
-            animateLoading = false,
-            asClickable = true,
-            asFilled = !this.asFilled
-        )
-    }
-}
-
-/**
- * Represents the error states that the actions section can assume.
- */
-internal sealed class ErrorState {
-    object None : ErrorState()
-    object UserNotLogged : ErrorState()
-    object UnknownError : ErrorState()
 }
